==> _setup/3a_setup_spp_threat_class_narr.Rmd <==
---
title: 'Spp threats by narrative and category'
author: "*Compiled on `r date()` by `r Sys.info()['user']`*"
output: 
  html_document:
    code_folding: hide
    toc: true
    toc_depth: 3
    toc_float: yes
    number_sections: true
    theme: cerulean
    highlight: haddock
    includes: 
      in_header: '~/github/src/templates/ohara_hdr.html'
  pdf_document:
    toc: true
---

``` {r setup, echo = TRUE, message = FALSE, warning = FALSE}

knitr::opts_chunk$set(fig.width = 6, fig.height = 4, fig.path = 'figs/',
                      echo = TRUE, message = FALSE, warning = FALSE)

source('https://raw.githubusercontent.com/oharac/src/master/R/common.R')
source(here::here('common_fxns.R'))
source(here::here('_setup/api_fxns.R'))
```

# Summary

Examine threats to marine species as determined by IUCN threat classifications and text dismemberment of IUCN species narratives.

# Data Sources

### IUCN Red List

# Methods

## Get IUCN threat lists for all available IUCN marine spp

``` {r get_threats_from_api}

spp_threats_file <- file.path(dir_bd_anx, 'iucn/threats',
                              sprintf('iucn_spp_threats_%s.csv', api_version))

spp_info <- read_csv(here('_data', 
                          sprintf('spp_marine_from_api_%s.csv', api_version)))

if(!file.exists(spp_threats_file)) {
  
  spp_ids <- spp_info$iucn_sid
  
  threat_url <- 'http://apiv3.iucnredlist.org/api/v3/threats/species/id/%s?token=%s'
  
  chunk_size <- 500
  n_chunks <- ceiling(length(spp_ids)/chunk_size)
  
  for(j in 1:n_chunks) { 
    ### j <- 1
    spp_index <- c( ((j - 1) * chunk_size + 1) : min((j * chunk_size), length(spp_ids)) )
    chunk_file <- file.path(dir_bd_anx, 'tmp', 
                            sprintf('spp_threat_chunk_%s_%s_%s.csv', 
                            min(spp_index), max(spp_index), api_version))

    if(!file.exists(chunk_file)) {
      message('Getting threat info for species ', min(spp_index), ' to ', max(spp_index))
      
      spp_ids_chunk <- spp_ids[spp_index]
      spp_threat_chunk <- mc_get_from_api(threat_url, spp_ids_chunk, api_key, 
                                          delay = .5, cores = 16, 
                                          verbose = FALSE)
      message('... found ', nrow(spp_threat_chunk), ' threat rows for these species')
      
      write_csv(spp_threat_chunk, chunk_file)
      
    } else {
      
      message('Chunk file ', basename(chunk_file), ' already exists; skipping these spp')
      
    }
  }
  
  spp_threat_chunk_files <- list.files(file.path(dir_bd_anx, 'tmp'), 
                                  pattern = 'spp_threat_chunk', 
                                  full.names = TRUE)
  
  spp_threats_df <- lapply(spp_threat_chunk_files, FUN = function(x) {
    read_csv(x) %>%
      mutate(code = as.character(code))}) %>%
    bind_rows() %>%
    rename(iucn_sid = id) %>%
    mutate(iucn_sid = ifelse(is.na(iucn_sid), param_id, iucn_sid)) %>%
    arrange(iucn_sid)
  
  
  spp_threats_df1 <- spp_threats_df %>%
    mutate(score_num = as.integer(str_extract(score, '[0-9]+'))) %>%
    select(-param_id, -api_error)
  
  write_csv(spp_threats_df1, spp_threats_file)
}

```

``` {r}

spp_threats <- read_csv(file.path(spp_threats_file)) %>%
  distinct() %>%
  mutate(impact = str_replace(score, ':.*', ''))
  
threat_class <- read_csv(here('_raw/iucn_threat_class.csv'))

spp_threat_cats <- spp_threats %>%
  group_by(code, timing, impact) %>%
  summarize(n_spp = n()) %>%
  ungroup() %>%
  left_join(threat_class, by = 'code') %>%
  mutate(cat = str_extract(desc, '.+?(?=:)'),
         cat_code = str_extract(code, '^[0-9]+') %>% as.integer()) %>%
  arrange(cat_code) %>%
  filter(!is.na(cat_code))
  
cats <- spp_threat_cats$cat %>% unique()

for(threat_cat in cats) {
  ### threat_cat <- cats[1]
  thr_cat_df <- spp_threat_cats %>%
    filter(cat == threat_cat) %>%
    mutate(desc = str_replace(desc, '^.+: ', ''))
  
  cat_plot <- ggplot(thr_cat_df, aes(x = desc, y = n_spp, fill = impact)) +
    ggtheme_plot() +
    theme(axis.title.y = element_blank()) +
    geom_bar(stat = 'identity') +
    coord_flip() +
    labs(title = threat_cat,
         y = 'number of affected species')
  
  print(cat_plot)
}
```

## Get IUCN narratives for all available IUCN marine spp

From the full IUCN species list, send each IUCN species ID into the API to get the narrative listed for that species. 

``` {r determine_spp_narrs}

spp_info <- read_csv(here('_data', 
                          sprintf('iucn_risk_current_%s.csv', api_version)))

spp_narr_file <- file.path(dir_bd_anx, 
                           sprintf('iucn/spp_narr_from_api_%s.csv', api_version))
reload <- FALSE

if(!file.exists(spp_narr_file) | reload) {
  
  message('Using API to determine species narrative from marine species info list')
  
  spp_ids_all <- spp_info %>%
    .$iucn_sid
  
  spp_narr_url <- 'http://apiv3.iucnredlist.org/api/v3/species/narrative/id/%s?token=%s'
  
  chunk_size <- 500
  n_chunks <- ceiling(length(spp_ids_all)/chunk_size)
  
  if(!dir.exists(file.path(dir_bd_anx, 'tmp'))) {
    dir.create(file.path(dir_bd_anx, 'tmp'))
  } ### unlink(file.path(dir_bd_anx, 'tmp/*narr*.*'))

  
  for(j in 1:n_chunks) { 
    ### j <- 2
    spp_index <- c( ((j - 1) * chunk_size + 1) : min((j * chunk_size), length(spp_ids_all)) )
    chunk_file <- file.path(dir_bd_anx, 'tmp', 
                    sprintf('spp_narr_chunk_%s_%s_%s.csv', 
                            min(spp_index), max(spp_index), api_version))

    if(!file.exists(chunk_file)) {
      message('Getting narrative info for species ', 
              min(spp_index), ' to ', max(spp_index))
      
      spp_ids_chunk <- spp_ids_all[spp_index]
      spp_narr_chunk <- mc_get_from_api(spp_narr_url, spp_ids_chunk, 
                                        api_key, cores = 12, delay = .5)
      message('... found ', nrow(spp_narr_chunk), 
              ' narrative rows for these species')
      
      write_csv(spp_narr_chunk, chunk_file)
      
    } else {
      
      message('Chunk file ', chunk_file, 
              ' already exists; skipping these spp')
      
    }
  }
  
  ### fields: 
  ### id | code | habitat | suitability | season | majorimportance

  spp_narr_chunk_files <- list.files(file.path(dir_bd_anx, 'tmp'), 
                                    pattern = 'spp_narr_chunk', 
                                    full.names = TRUE)
  
  spp_narr_df <- lapply(spp_narr_chunk_files, FUN = function(x) read_csv(x)) %>%
    bind_rows() %>%
    rename(iucn_sid = species_id) %>%
    mutate(iucn_sid = ifelse(is.na(iucn_sid), param_id, iucn_sid)) %>%
    arrange(iucn_sid) %>%
    select(-name) %>%
    distinct()
  
  # spp_errors <- spp_narr_df %>%
  #   filter(!is.na(api_error) & api_error != 'no data.frame') %>%
  #   .$iucn_sid
  ### all these errors are due to returning a zero-length list instead of a data.frame

  write_csv(spp_narr_df, spp_narr_file)
  
} else {
  
  message('File of species narratives from API exists: \n  ', spp_narr_file)

}

```
