==> _setup/3b_iucn_threat_class_table.Rmd <==
---
title: "IUCN threat classification scheme"
author: '*Compiled on `r date()` by `r Sys.info()["user"]`*'
output:
  word_document:
    toc: no
    toc_depth: '3'
  pdf_document:
    toc: yes
  html_document:
    code_folding: hide
    highlight: haddock
    includes:
      in_header: ~/github/src/templates/ohara_hdr.html
    number_sections: yes
    theme: cerulean
    toc: yes
    toc_depth: 3
    toc_float: yes
---

``` {r setup, echo = FALSE, message = FALSE, warning = FALSE}

knitr::opts_chunk$set(fig.width = 6, fig.height = 4, fig.path = 'figs/',
                      echo = TRUE, message = FALSE, warning = FALSE)

source('https://raw.githubusercontent.com/oharac/src/master/R/common.R')
source(here::here('common_fxns.R'))

```

# Summary

Document the threat classification scheme as a lookup table

# Data Sources

https://www.iucnredlist.org/resources/threat-classification-scheme, accessed 1/26/2019

# Methods

## Threat Impact

For each threat, it is recommended that the timing of the threat (i.e. past, ongoing or future), its scope (i.e. the proportion of the total population affected) and severity (i.e the overall declines caused by the threat) should be recorded.

Timing options are:

* Only in the past and unlikely to return
* In the past but now suspended and likely to return
* Ongoing
* Only in the future
* Unknown
 

Scope options are:

* Affects the whole population (>90%)
* Affects the majority of the population (50-90%)
* Affects the minority of the population (<50%)
* Unknown
 

Severity options are:

* Causing or likely to cause very rapid declines (>30% over 10 years or three generations; whichever is the longer)
* Causing or likely to cause rapid declines (20–30% over 10 years or three generations; whichever is the longer)
* Causing or likely to cause relatively slow but significant declines (<20% over 10 years or three generations; whichever is the longer)
* Causing or likely to cause fluctuations
* Causing or likely to cause negligible declines
* No declines
* Unknown
 

Any analysis of the threats should preferably take into account the timing, scope and severity of the threats to calculate an impact score (see attached sheet for calculating threat impact scores) and should also consider how the threats impact the taxa concerned, as recorded by the stresses. These additional attributes are displayed on the Red List website.

## Classification scheme

The full hierarchical structure of threat types displayed in IUCN Red List assessments is shown below.

Direct threats are the proximate human activities or processes that have impacted, are impacting, or may impact the status of the taxon being assessed (e.g., unsustainable fishing or logging, agriculture, housing developments, etc.) Direct threats are synonymous with sources of stress and proximate pressures.

In using this hierarchical classification of the drivers of species decline, Assessors are asked to indicate the threats that triggered the listing of the taxon concerned at the lowest level possible. These threats may have occured in the past ("historical, unlikely to return" or "historical, likely to return"); or may be "ongoing"; or may be likely to occur in the "future", using a time frame of three generations or ten years (whichever is the longer, up to a maximum of 100 years in the future), as required by the IUCN Red List Criteria. The 'Major Threats' referred to in the 'Required and Recommended Supporting Information for IUCN Red List Assessments', are threats coded as having High or Medium impacts (see threat impact scoring below).

The attached working document provides a list of the threat types with definitions, examples of the threats and guidance notes on using the scheme. Comments on the Threats Classification Scheme are welcome.

```{r}

threat_class <- read_csv(here('_raw/iucn_threat_class_raw.csv')) %>%
  separate(code, c('l1', 'l2', 'l3'), remove = FALSE) %>%
  mutate(cat = description, subcat = description) %>%
  mutate(cat = ifelse(is.na(l2), cat, NA),
         subcat = ifelse(is.na(l3), subcat, NA)) %>%
  fill(cat, subcat, .direction = 'down') %>%
  mutate(desc = case_when(is.na(l2) ~ cat,
                          is.na(l3) ~ paste0(cat, ': ', subcat),
                          TRUE      ~ paste0(cat, ': ', subcat, ': ', description))) %>%
  select(code, desc) %>%
  distinct()
  
write_csv(threat_class, here('_raw/iucn_threat_class.csv'))

```

```{r}

spp_threats <- read_csv(file.path(dir_bd_anx, 'iucn/threats',
                                  sprintf('iucn_spp_threats_%s.csv', api_version))) %>%
  filter(!is.na(iucn_sid) & !is.na(code)) %>%
  # filter(timing != 'Past, Unlikely to Return' | is.na(timing)) %>% ### ~ 635
  # filter(score_num >= 3 | is.na(score)) %>%
  .$code %>% unique()

threat_class_marine <- read_csv(file.path(dir_git, '_raw/iucn_threat_class.csv')) %>%
  filter(code %in% spp_threats | !str_detect(code, '\\.')) %>%
  mutate(code = fct_inorder(code))

library(flextable)

ftable_theme <- function(df) {
  flextable(df) %>%
    theme_zebra() %>%
    align(part = 'all', align = 'center') %>%
    bg(bg = 'white', part = 'header') %>%
    border(border.bottom = officer::fp_border(), part = 'header') %>%
    font(font = 'Arial', part = 'all') %>%
    fontsize(size = 8, part = 'all')
  
}

threat_table <- ftable_theme(threat_class_marine) %>%
  width(j = 2, 5)

threat_table
```
