==> _setup/5d_generate_taxon_list.Rmd <==
---
title: 'Generate a list of taxonomic ranks'
author: "*Compiled on `r date()` by `r Sys.info()['user']`*"
output: 
  html_document:
    code_folding: hide
    toc: true
    toc_depth: 3
    toc_float: yes
    number_sections: true
    theme: cerulean
    highlight: haddock
    includes: 
      in_header: '~/github/src/templates/ohara_hdr.html'
  pdf_document:
    toc: true
---

``` {r setup, echo = TRUE, message = FALSE, warning = FALSE}

knitr::opts_chunk$set(fig.width = 6, fig.height = 4, fig.path = 'figs/',
                      echo = TRUE, message = FALSE, warning = FALSE)

source('https://raw.githubusercontent.com/oharac/src/master/R/common.R')

source(here('common_fxns.R'))

```


# Summary

Generate a lookup of IUCN species IDs to taxonomic groups (the shapefile groups as well as the taxonomic ranks) to be used for plotting by taxon.

# Methods

```{r}
maps_file <- here('_data', sprintf('spp_marine_maps_%s.csv', api_version))
comp_file <- here('_data', sprintf('iucn_comp_assessed_%s.csv', api_version))
info_file <- file.path(dir_bd_anx, 
               sprintf('iucn/spp_info_from_api_%s.csv', api_version))
# bird_file <- here('_raw', 'seabird_species_2019.csv')

spp_maps <- read_csv(maps_file, col_types = cols('iucn_sid' = 'i',
                                                 'subpop'   = 'c')) %>%
  filter(presence != 5) %>%
  select(iucn_sid) %>% ### just need ID numbers here
  distinct()

spp_comp <- read_csv(comp_file, col_types = cols('iucn_sid' = 'i')) %>%
  select(-sciname, assess_gp)

spp_incl <- spp_maps %>%
  left_join(spp_comp, by = c('iucn_sid')) %>%
  filter(!is.na(assess_gp)) %>%
    ### drop non-comp-assessed
  distinct()
  
rank_df <- read_csv(info_file) %>%
  janitor::clean_names() %>%
  filter(iucn_sid %in% spp_incl$iucn_sid) %>%
  select(iucn_sid, kingdom:family) %>%
  mutate_if(is.character, .fun = function(x) tools::toTitleCase(tolower(x)))

spp_ranks <- spp_incl %>%
  left_join(rank_df, by = 'iucn_sid') %>%
  mutate(rank = case_when(phylum == 'Chordata' ~ class, 
                          kingdom == 'Plantae' ~ kingdom,
                          TRUE ~ phylum)) %>%
  distinct()

write_csv(spp_ranks, here('int/spp_ranks.csv'))

sum_df <- spp_ranks %>%
  group_by(kingdom, phylum, class, assess_gp) %>%
  summarize(n_spp = n_distinct(iucn_sid))

knitr::kable(sum_df)
```
